// szteam-network-frontend/src/app/services/models/page-response-game-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GameResponse } from '../models/game-response';
export interface PageResponseGameResponse {
  content?: Array<GameResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}